# Improve texture shader performance on fragments away from corners