@peppy 2023-04-08
On macOS (m2), performance with 100 of the test boxes jumps from 60fps to 72fps (16.7ms -> 13.6ms per draw frame). A pretty massive improvement.